+++ app/src/androidTest/java/woowacourse/shopping/feature/main/MainActivityTest.kt
@@ -18,15 +18,15 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.shopping.R
 import woowacourse.shopping.RecyclerViewItemCountAssertion
-import woowacourse.shopping.data.product.ProductDummyRepository
 import woowacourse.shopping.data.product.ProductRepository
+import woowacourse.shopping.data.product.remote.ProductRemoteRepository
 import woowacourse.shopping.imageUrl
 import woowacourse.shopping.price
 import woowacourse.shopping.title
 
 @RunWith(AndroidJUnit4::class)
 class MainActivityTest {
-    private val productRepository: ProductRepository = ProductDummyRepository
+    private val productRepository: ProductRepository = ProductRemoteRepository()
 
     @Before
     fun setUp() {
