+++ app/src/main/java/woowacourse/shopping/data/product/remote/UrlPatterns.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.product
+package woowacourse.shopping.data.product.remote
 
 const val PRODUCT_FIND_PATH = "/products/find/"
 const val PRODUCT_FIND_ALL_PATH = "/products"
