+++ app/src/main/java/woowacourse/shopping/feature/detail/ProductDetailActivity.kt
@@ -8,12 +8,12 @@ import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.snackbar.Snackbar
 import woowacourse.shopping.R
 import woowacourse.shopping.common.QuantityControlListener
-import woowacourse.shopping.data.cart.CartDatabase
-import woowacourse.shopping.data.cart.CartDummyRepository
-import woowacourse.shopping.data.inquiryhistory.InquiryHistoryDatabase
-import woowacourse.shopping.data.inquiryhistory.InquiryHistoryLocalRepository
-import woowacourse.shopping.data.product.ProductClient
-import woowacourse.shopping.data.product.ProductRemoteRepository
+import woowacourse.shopping.data.cart.local.CartDatabase
+import woowacourse.shopping.data.cart.local.CartDummyRepository
+import woowacourse.shopping.data.inquiryhistory.local.InquiryHistoryDatabase
+import woowacourse.shopping.data.inquiryhistory.local.InquiryHistoryLocalRepository
+import woowacourse.shopping.data.product.remote.ProductClient
+import woowacourse.shopping.data.product.remote.ProductRemoteRepository
 import woowacourse.shopping.databinding.ActivityProductDetailBinding
 import woowacourse.shopping.feature.cart.CartActivity
 import woowacourse.shopping.feature.detail.viewmodel.ProductDetailViewModel
