+++ app/src/main/java/woowacourse/shopping/feature/main/MainActivity.kt
@@ -8,12 +8,12 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import woowacourse.shopping.data.cart.CartDatabase
-import woowacourse.shopping.data.cart.CartDummyRepository
-import woowacourse.shopping.data.inquiryhistory.InquiryHistoryDatabase
-import woowacourse.shopping.data.inquiryhistory.InquiryHistoryLocalRepository
-import woowacourse.shopping.data.product.ProductClient
-import woowacourse.shopping.data.product.ProductRemoteRepository
+import woowacourse.shopping.data.cart.local.CartDatabase
+import woowacourse.shopping.data.cart.local.CartDummyRepository
+import woowacourse.shopping.data.inquiryhistory.local.InquiryHistoryDatabase
+import woowacourse.shopping.data.inquiryhistory.local.InquiryHistoryLocalRepository
+import woowacourse.shopping.data.product.remote.ProductClient
+import woowacourse.shopping.data.product.remote.ProductRemoteRepository
 import woowacourse.shopping.databinding.ActivityMainBinding
 import woowacourse.shopping.feature.cart.CartActivity
 import woowacourse.shopping.feature.detail.ProductDetailActivity
