+++ app/src/test/java/woowacourse/shopping/feature/main/viewmodel/MainViewModelTest.kt
@@ -6,14 +6,14 @@ import org.junit.jupiter.api.Assertions.assertAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
-import woowacourse.shopping.data.cart.CartDao
-import woowacourse.shopping.data.cart.CartDummyRepository
 import woowacourse.shopping.data.cart.CartRepository
-import woowacourse.shopping.data.inquiryhistory.InquiryHistoryDao
-import woowacourse.shopping.data.inquiryhistory.InquiryHistoryLocalRepository
+import woowacourse.shopping.data.cart.local.CartDao
+import woowacourse.shopping.data.cart.local.CartDummyRepository
 import woowacourse.shopping.data.inquiryhistory.InquiryHistoryRepository
+import woowacourse.shopping.data.inquiryhistory.local.InquiryHistoryDao
+import woowacourse.shopping.data.inquiryhistory.local.InquiryHistoryLocalRepository
 import woowacourse.shopping.data.product.ProductDummyRepository
-import woowacourse.shopping.data.product.ProductRepository
+import woowacourse.shopping.data.product.remote.ProductService
 import woowacourse.shopping.feature.InstantTaskExecutorExtension
 import woowacourse.shopping.feature.getOrAwaitValue
 import woowacourse.shopping.imageUrl
@@ -25,7 +25,7 @@ class MainViewModelTest {
     private lateinit var viewModel: MainViewModel
     private val cartDao = mockk<CartDao>()
     private val inquiryHistoryDao = mockk<InquiryHistoryDao>()
-    private val productRepository: ProductRepository = ProductDummyRepository
+    private val productRepository: ProductService = ProductDummyRepository
     private val cartRepository: CartRepository = CartDummyRepository(cartDao)
     private val inquiryHistoryRepository: InquiryHistoryRepository = InquiryHistoryLocalRepository(inquiryHistoryDao)
     private val pageSize: Int = 20
